Preface: initial submission
Add the preface and preprocess it with sed, trying to sweep as many
LaTeX macros as we can. This will immensely help us with cleanup.

I'm going to keep the sed spells around for the next chapaters.
They can be applied with

    sed -E -f autofix.sed -i '' ${file.md}

There are some questionable decisions about the labels and index
markers, but I guess I'll go with the current approach.