hacks: Use constants where you mean constants
I was not that well versed with Pollen syntax back when I started
writing and using those. Make sure to use actual constants where
a constant is needed. This enables better syntax highlighting.

There are a bunch of cases where vertical bars are necessary, such as
when the constant is used right before a period (as Racket syntax for
identifiers allows periods). I mean, this *is* solvable (by duplicating
constants), but fuck that.