hacks: Chapter 4: single-page rendering
Start hacking around a way to render entire chapters onto a single page.
This is some kind of a way I'd like to use: fetch all relevant pagenodes
then extract their document data, then concatenate it, and boom, done.
It's probably not the most efficient way, but hey, it seems to work.

I'd need this form for multiple uses. First of all, this is probably
what's going to be shipped to third-party editors if I ever use them.
(Though, it probably will be copied into Microsoft Word, and I'd need
to add some more tweaks for that too.)

Second, it would probably be a nice way to present the book too, with
just one chapter on a page. For those users who'd like it that way.
Also, it will be nice to get *everything* onto a single page too.

But for now, just save this piece of code.