pollen: Chapter 5: section 5.2.1
So here's the plan: I write denotation is Scheme, the way they are coded
in the book, and then at some later point in time I will write some tool
to convert them into denotation. Christian used his own tool (Lisp2TeX)
for this, so I guess it's time for reinventing the wheel, yay! But not
today. Today I want to make progress on the book and don't want to be
like a monkey translating all those LaTeX into broken LaTeX.