hacks: Rendering math with KaTeX
Integrate (somewhat) KaTeX into the publishing pipeline.

KaTeX is a Node.js library. While rendering itself is somewhat fast,
starting up Node.js from Racket is not, so I keep around a simple
rendering server, making pages render much faster (but still slow).

I don't invest too much into this being a "nice" pipeline to use.
You'd need to "npm ci" manually in book-src to download KaTeX,
but after that simple "raco pollen start" works fine for me.

Note that a lot of math on the pages is not actually renderable right
now. This is going to be dealt with in the following commits.